backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
pyyaml>=6.0.2
loguru>=0.7.2

# 事件总线
pyee>=12.0.0

# 数据库
aiosqlite>=0.20.0
chromadb>=0.5.0

# ASR
pyaudio>=0.2.14
webrtcvad>=2.0.10
funasr>=1.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Embedding
sentence-transformers>=3.3.0

# LLM
openai>=1.54.0
anthropic>=0.40.0

# TTS
websockets>=13.0

# 音频播放
pygame>=2.6.0

# HTTP 客户端
httpx>=0.27.0

# 文件上传（multipart/form-data）
python-multipart>=0.0.9
